+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 19-FEB-2026 CEST
+Last change: 20-FEB-2026 CEST
 
 0.9.7
 -----
@@ -10,6 +10,7 @@ Last change: 19-FEB-2026 CEST
 - Fixed labels of the vertical lines in the subcluster test plot
 - Fixed automatic Localize loading/unloading z-calibration paths when changing cameras
 - Fixed ``rel_sigma_z`` in G5M (previously incorrectly divided by pixel size)
+- Fixed loading square picks in Render
 
 0.9.6
 -----